Add files via upload
The development of a Heart Attack Prediction GUI using Machine Learning and Python 
represents a significant stride towards harnessing advanced technologies for proactive 
healthcare. This project amalgamates the power of machine learning algorithms with the 
versatility of Python programming to create a user-friendly interface capable of predicting the 
likelihood of a heart attack. Through the utilization of robust datasets and sophisticated 
predictive models, the GUI provides a valuable tool for early detection and prevention of 
cardiovascular events.